Fix bug in Rusanov flux in parallel
Found a bug in the Rusanov flux. I put that in when I optimized it
recently. The symmetric part of the flux was using the weighted normal
that was only supposed to be used for the dissipation term.

The (old) code below is wrong, vnl and vnr are never used, but
overwritten by what is under dissipation and that is what is used for
both parts of the flux.

    // face-normal velocities
    real vnl = ul*nx + vl*ny + wl*nz;
    real vnr = ur*nx + vr*ny + wr*nz;

    // dissipation
    auto len = tk::length( {mx,my,mz} );
    vnl = ul*mx + vl*my + wl*mz;
    vnr = ur*mx + vr*my + wr*mz;
    auto sl = std::abs(vnl) + al*len;
    auto sr = std::abs(vnr) + ar*len;
    auto smax = std::max( sl, sr );

    // Numerical fluxes
    fr  = 0.5*(rL*vnl + rR*vnr - smax*(rR - rL));
    fru = 0.5*(ruL*vnl + pl*nx + ruR*vnr + pr*nx - smax*(ruR - ruL));
    frv = 0.5*(rvL*vnl + pl*ny + rvR*vnr + pr*ny - smax*(rvR - rvL));
    frw = 0.5*(rwL*vnl + pl*nz + rwR*vnr + pr*nz - smax*(rwR - rwL));
    fre = 0.5*((reL + pl)*vnl + (reR + pr)*vnr - smax*(reR - reL));

This commit fixes that issue.