Only derefine active tets when uniform deref
Previously all edges were marked for derefinement when uniform
derefinement was done. Now only the active tets' edges are marked for
derefinement, which avoids marking edges of grandparents, so it does the
right thing if two uniform refinement steps are followed by a uniform
derefinement step.

Thanks to @rfbird for the help with this.