Add outref/outderef logic to Refiner
This commit puts in logic to the end of DG::writeFields() to call into
Refiner and perform a mesh refinement step so that we can output
solution data on a refined mesh. We call this mesh "field-output mesh".

The idea is to perform a uniform refinement across the entire mesh,
evaluate and write out the solution on the refined mesh, and derefine
the mesh to the state before the write.

This commit puts in only the logic for the refinement, followed by the
derefinement, followed by a write using the refined (field-output) mesh.

Although, we do not yet actually refine and derefine the mesh and also
we do not yet evaluate the solution on the field-output mesh, I have
done some initial testing saving the refined (field-output) mesh (but
not yet the solution).

The logic so far appears to play along well with initial mesh refinement
(t0ref) and during-timestepping mesh refinement (dtref), to the extend
tested by the existing regression tests.