Reorganize field output logic in DG to enable outref
This commit reorganizes the logic around field output in DG to enable
both the required communication step for nodal output as well as field
output on a refined mesh. Both nodal output and refined field output are
optional.

It also puts in necessary code for copying over the higer degrees of
freedom for evaluating the P1 solution on the refined field output mesh,
in DG::solution().