Put in transformation of DG(P1) slopes for outref
This commit puts in code in DG::solution() to transform the high-order
degrees of freedom for DG(P1) from Dubiner basis to slopes and back, so
that the solution in nodes can be properly evaluated on a mesh refined
for field output.

This is so far only tested in serial and for Transport, and not
regression tested.