Move output field NDOF out of problem class
Putting in the name of an extra field output, such as 'ndof' for pDG is
better done in DG instead of in GaussHumpCompflow, because DG is the one
that decides to add the field for the field output.

This will be revamped when we put in a proper class responsible for
handling field output based on user input.