Make DG have its own copy of the mesh
DG augments the mesh connectviity and coordinates. Previously it
extended the one in Discretization, but it with mesh refinement (also
used for field output), it becamse unwieldly to reason about code if the
"master" copy in Discretization is changed. So this commit makes DG have
its own copy of inpoel and coords so it can screw with them all it
wants.

Also removed a bunch of unsafe non-const-ref accessors from
Discretization.