Change DG node comm exit condition to use Disc::nodeCommMap
This commit switches the exit condition, i.e., when a chare is finished
received all contributions to its node field output, to use the node
communication map in Discretization, instead that of the potentially
refined mesh.

This, I think, should fix a bug in the asynchronous logic in parallel,
which allows a chare to receive contributions from other chares before
initializing its own m_outmesh.nodeCommMap. If that happens, the exit
condition is wrong and either prematurely exists from the communication
or deadlocks. The node comm map in Discretization should already be
final for the coarse mesh by the time the node field communication uses
it.

I believe this solution should work because the node comm map in
Discretization while stores a different node-list data if we refine the
mesh for field output, its keys should be the same for both the original
and the refined mesh, since the communication pattern (i.e., which
chares communicate with each other) should not change with mesh
refinement.

It goes without saying that this bug only appears randomly, and I cannot
reproduce locally - at this point this is a guess at a fix.