Put in initial parsing for outvar ... end blocks
Works from both field_outpout and history_output.

So far can parse the following:

    outvar
      density
      momentum
      total_energy
      velocity
      pressure
      U1 U2
    end

In the above example, besides the depvar-based U1 and U2, only the
keywords density, momentum, total_energy, velocity, and pressure are are
the ones accepted, and everything else triggers a parser error,
including "pressuree". This is facilitated by tk::grm::exact_scan. If
tk::grm::scan is used in inciter::deck::outvar instead of
tk::grm::exact_scan, "pressuree", and "pressureasdfasdf", etc., are
accepted as correct keywords, which is incorrect. I suspect
tk::grm::exact_scan might also work for fixing #341, but I am not ready
to open up the can of worms that would come with simply replacing
tk::grm::scan with tk::grm::exact_scan. The latter is only used from
inciter::deck::outvar for now.

The depvar-based U1 and U2 are accepted. The lower case u1 and u2 are
incorrect, because these assume that they are instantaneous (in a
turbulence sense) variables and requesting fluctuating variables make no
sense without defining the mean first.

U must be a depvar upstream. If not, it is correctly not accepted. E.g,
C1 (or c1) will not be accepted if

  compflow
    depvar u
  end

is configured upstream.

For now U6 and larger is incorrectly accepted. I.e., the existence of
"compflow" with depvar 'u' should trigger an error when referring to a
component larger than 5.