Parse output variables as OutVar
The new helper class, OutVar, defined in Control/OutVar.hpp is similar
to Term, defined in Control/StatCtr.hpp.

Inciter's grammar now parses the output variables in a var ... end block
not as simple strings but as a list of OutVar.

So far we can parse two types of outvars: human-readabale and plain. The
following example has both types:

  field_output
    var
      density momentum  # important
      total_energy      # not that important
      velocity
      pressure
      U1 U2             # crucial
    end
    interval 5
    sideset 1 2 3 end
  end

The above is successfully parsed into a std::vector< tk::ctr::OutVar >,
stored behind g_inputdeck.get< tag::cmd, tag::io, tag::outvar >(). The
comments above just signify that comments can be added between the
variable specifications.

OutVar stores, in a typesafe fashion, the following fields:

  char var;             //!< Variable name
  ncomp_t field;        //!< Field ID
  Centering centering;  //!< Centering
  std::string name;     //!< Human readable name

For human-readable outvars 'var' and 'field' are unused and contain '0'
and 0, respectively, and 'name' is non-empty. For plain outvars 'var'
contains the depvar, which identifies the eq system, and 'field' is the
field id parsed, which is still not bound-checked. The (internal) type
of outvar also determines how we output an OutVar to screen, see
operator<< in Control/OutVar.hpp. Output for the above var ... end
block:

 * Output fields:
 ----------------
     Field(s)                                 : { density_n momentum_n pressure_n total_energy_n velocity_n U1n U2n }
     Surface side set(s)                      : { 1 2 3 }

The default centering for all outvars is Centering::NODE.

This commit also contains a fix to exact_scan, that previously
erroneously switched the argument of pegtl::until< R, S... >.