Put in grammar for recognizing centering keywords in inciter's outvar blocks
With this commit, inciter's input file parser now recognizes the 'node'
and 'elem' keywords in all var ... end blocks. The centering is saved in
the parser's state so that any variable specification following the
given centering keyword is understood as the given centering.

Also remove the postfixed centering from the screen echo of all outvar
names. This is to make it consistent with the input file lists.

Also screen echo now categorizes node and elem outvar lists, which is
cleaner and eliminates the need for a centering postfix to the variable
names.