Add alias to output variable parsing
Bow the following can be parsed:

  field_output
    var
     elem analytic "c0_analytic"
     node C1 "c0_numerical"
    end
    interval 5
    #refined true
  end

The above assigns the quoted strings to the preceding output variable
request. Aliases are optional and if exist, they will be used to label
output variables in output files.

This is done by adding a new field to OutVar::alias. OutVar::name is a
human-readable name that inciter will know about, i.e., "pressure",
"density", which will have meaning inside the code and so we will map
them to physics variables. OutVar::alias is only known by the user and
will only be used for output.

Also add new built-in human-readable variable 'analytic'.