Skip trying to extract human-readable field output variable
This simply avoids a runtime error. Next is to cook up a mechanism that
allows calling a custom function that can extract an arbitrary variable
from a soution vector, preferably independent of the eq system type and
the variable type.

Also add more docs on the difference between depvar-based and
human-readable outvars where they are constructed.