Add configurable functions for human-readable outvars
The main contribution of this commit is to add a machinery for
specifying and configuring functions that extract/compute physics
variables, defined as human-readable output variables that are known to
the code based on physics, from the numerical solution. Example:

compflow
  ...
end

field_output
  density
end

As a result of the above input file block the code is now smart enough
to recognize that a compflow ... end block was configured upstream of
the field_output ... end block from which the output variable 'density'
is requested by the user. Internally, 'density' is a human-readable
output variable. Human-readable output variables have not
OutVar::name.empty().

The way this works is that there is a list of human-readable output
variables that will be known to the code. For now this is only
CompFlow::density. How to compute these variables from the numerical
solution is known by the particular PDE system, so a function is
assigned to the output variable, that if requested by the user in the
input file, the function that computes the variable is assigned to the
OutVar so the correct output can be generated for field output. This
will allow for adding arbitrary kinds of field output for arbitrary PDE
types. Also, it will be possible to configure different functions for
different PDE types. E.q., velocity may be computed differently by
MultiMat compared to CompFlow. Also, this will be used for all kinds of
field output, including surface and history output.

Other changes and background:

* OutVar.hpp now lives under Control/Inciter/OutVar.hpp. Also, OutVar is
now defined in namespace inciter::ctr instead of tk::ctr. These changes
signify that OutVar is now specific to inciter and not intended (at this
point) for other executables, e.g., walker. This is because the new
field/state, getvar, a std::function, is initialized by calling into an
inciter-specific set of functions, interrogating PDEs (a unique feature
of inciter).

* Note that the new field, OutVar::getvar is a std::function and by
default that is not PUPable. But we were still able keep the list of
OutVars in inciter's inputdeck because we can initialize the
std::function when Charm++ unpacks it by running through the
free-function, inciter::assignGetVar(), declared and defined in
PDE/ConfigureOutVar.[ch]pp, which is the main entry point for assigning
all such functions for inciter. (Only 'density' is configured for now.)

* Added a lot more comments to the part of the grammar where OutVars are
instantiated, explaining how and why.

* Fixed the erroneous grammar for outvar alias parsing.

* Outputting variables to file now use the alias as labels, if
configured.

* Human-readable outvars are now correctly written to file - only
'density' is configured for now.

* If multiple compflow ... end blocks are configured by the user, the
names of human-readable outvars (OutVar::name) are postfixed by '_' +
depvar. Their getvar() should still work in that case since we search
for the built-in/known name, e.g., 'density' in the output variable name
requested, so 'density_2', should match a getvar for density and the
getvar functions are parameterized with the eq system offset, so it
should output the variable from the correct eq system.

* If multiple compflow ... end blocks are configured, and a
human-readable variable is requested, at this time, the variable from
all eq systems configured are written out to file. (Obviously, it makes
not much sense to configure multiple compflow ... end blocks, but it
does for transport ... end, for which this should also work.)

Now we are in a close position to configure the existing regression
tests to use this machinery and not fail. That's probably next.