Add offset-system map to inciter
This is stored in inciter::g_inputdeck.get< tag::sys >(), and it is a
map that associates the equation system index (value) to variable
offsets (key) for all scalar variables for all equation systems
configured by the user and thus instantiated from the PDE factory. Note
that the system index in the map value is the system index within a
single equation type. This the same as in a PDE object's m_system.

This map allows looking up the system index based on the variable offset
(obviously), and this thus eliminates the requirement for the knowledge
of the system index if the variable offset is available. This in turn
allows outputting physics variables independent of the PDE classes. This
way variable output code stays generic and independent of the equation
type. Thus now we are ready to make an EOS call within a variable output
getvar function, e.g., pressure, without having to pass the equation
system index besides the variable offset.

This also means that we can probably get rid of the m_system in all PDE
objects since the offset is known there already and using the this map,
now the system index can be looked up. This should also simplify a bunch
of function signatures that so far have required passing down the system
index. But this change is for another day.

Implementation details:

Since the system map that this commit adds to inciter's input deck
should only store data for the instantiated PDE systems, the code to
generate this map is best available inside PDEStack::createPDE(), since
that is the function that instantiates all PDEs.