Port pDG tests to new outvar machinery
Had to hook back up PDE::solution() that calls initialize() to evaluate
the analytic solution (or set ICs at t=0) because DG's ElemDiagnostics
needs it to compute the distance between the numerical solution and the
ICs. This is a work-around so the pDG tests do not fail on their
diagnostics, which test on these quantities.

Eventually, the diagnostics variables should also be part of the new
outvar machinery, not only field, surface, and history output, and that
time Problem::solution() can go away.

With this commit all DG regression tests now have been ported to the new
outvar machinery, except multimat, so the following now all pass:

ctest -j32 -L dg -LE multimat