Put in parsing for multimat outvars
This is the equivalent of depvar-based outvars but different than those
for any other equation types, e.g., CompFlow or Transport. This is
because multimat outvars are configured based on an acceptable character
label (see inciter::deck::multimatvars), and a material index (instead
of a depvar + a component index, as for other equation types).

Example:

  multimat
    ...
    depvar c    # depvar is still required because it is used
                # internally, as well as used for identifying the
                # eq system in the input file
    nmat 2      # 2-material problem
  end

  field_output
    var elem D1 D2 W2 end
  end

Requesting the density of material 1 and 2, and the z velocity of
material 2 for field output.