Configure multimat for new outvar machinery
This commit configures the multimaterial eq to use the new output
variable machinery for field output. All regressions now must pass.

Details:

* Specifying the 'scheme' in the input file now must proceed any output
variable configuration block, because the material-based output
variables need the proper settings for NDOF/RDOF to compute the eq
scalar component indices based on material ids. The grammar has been
modified to set the NDOF/RDOF immediately after the scheme token is
parsed, instead of at the end of parsing the input file as was
previously. (This is facilitated by tk::grm::process_alpha, which allows
for invoking multiple grammar actions.)

* In inciter's grammar, multimatvars now assigns the Dof-version of the
MultiMat indexing functions.

* Multimat now errors if no depvar is configured. This is still
required, as itis used internally, and to label eq systems in the input
file, but not used for multimat outvars based on material id.

* This commit also introduces a new type of OutVar, that is configured
based on material id. Internally, inside OutVar, matvars
(material-id-based output variables), are represented very similarly to
depvar-based outvars, with a new addition of a non-empty 'matvar' field,
which is a string that stores how the user has configured the outvar in
the input file. OutVar::matvar is also used to decide whether the
conserved-variable solution vector or the primitive variable solution
vector is used to extract the output variable. See OutVar::primitive().

* inciter::assignGetVars() is now guarded against the runtime system
calling this function with an empty inputdeck. This allows detecting
which type of eq system is configured (compflow or multimat) and thus
allows assigning the correct getvar functions for human-readable
outvars, even if they keywords, e.g., density, would configure different
functions depending on the eq configured. We assumed that no compflow
AND multimat is configured by the user.