Revert "Correctly skip multimat for bounds checking"
This reverts commit 5d6a3b7d46eaaa57ee59805bda83b57b5ee309bd, because it
was correct after all. That's because check_outvar runs for each outvar
parsed, testing the last one just added to the list of outvars. If it is
a matvar, it should correctly skip the test (because it will be tested
in push_matvar), and if it is not a matvar, it was correct originally.