Enable multiple meshes in inciter
This commit enables computing on multiple meshes in inciter. Example
use with 3 meshes:

./charmrun +p12 inciter -v -c vortical_flow_alecg.q -i mesh1.exo -i mesh2.exo -i mesh3.exo -u 0.9

This is not yet tested very well, but appears to work fine in parallel,
with arbitrary number of meshes, with nonzero virtualization, in SMP and
non-SMP mode, with or without PE-local reordering, and with initial
uniform mesh refinement (which refines all meshes for now). Really only
tested with ALECG as the hydro scheme, but other schemes should also
work. If multiple meshes are given, for now we only instantiate the same
hydro scheme and compute the same problem on each mesh, but obviously,
the idea is to be able to instantiate a different physics scheme for a
different mesh and compute a coupled problem. There is no interaction
among the meshes/problems/schemes yet - that's next.

Thanks to @epmikida for suggesting this solution, which relies on
instantiating multiple chare arrays, one for each mesh. I believe this
will nicely hook up with ExaM2M for transferring the solution among
meshes.

These changes should be completely transparent to running with a single
mesh, i.e., all regression tests should pass.