Separate out compiler defs from QuinoaConfig
This allows including the new QuinoaBuildConfig.hpp, generated by cmake,
into Charm++ .ci files and pull in build config via compiler macros.
This separation is necessary because previously std lib headers were
required and that is not without trouble to push through charmc with the
correct include paths, etc. This allows compiler defs to be used in .ci
files.