Setup cross-solver callbacks
This commit adds code to setup cross-solver callbacks for solvers that
are coupled and thus involved in mesh-to-mesh transfer.

Problem and background:

The mesh-to-mesh transfer library, ExaM2M, currently requires a pair of
calls to setSourceTets() and a setDestPoints() to initiate a solution
transfer. The first is called on the source and the second is called on
the destination solver, passing the mesh (connectivity and node
coordinates) and a solution, defined on the mesh. The library then
transfer the solution from the source to the destination mesh. When the
transfer is complete, as the API is configured now, the library calls a
two callbacks, both passed to setDestPoints(). The idea is that both the
source and the destination solver provides a callback that are both
called so both can continue after the transfer is complete.

Solution:

To setup the two callbacks, the child solver (now only ALECG) is
configured to create the callbacks in its constructor, which then are
passed to its base (Discretization). This is because the child class is
in the best position to create callbacks (tied) to its own meshid, array
proxy, and proxy element. Since all solvers (Discretization bases,
holding different meshes) hold a list of transfer objects,
std::vector<Transfer>, every child solver has the same view of all,
potentially multiple, transfers prescribed. Thus the callbacks are
created and transferred to their base, Discretization, via
Discretization::transferCallback().

The callbacks are then processed further: (1) keep one to use for the
case when the solver is not coupled to any solver, (2) if the solver is
involved in a transfer as a source, the callback is sent to the
destination, and (3) if the solver is involved in a transfer as a
destination, the callback is stored in the list of transfers as one of
the callbacks. In the receive side of step (2) the destination solver
receives the callback from the source, Discretization::comcb(), and
stored as the second callback that will be called when the transfer is
complete by the library.

The communication of the callbacks among the coupled solvers are
complete, when all transfers that a solver is involved in as a
destination have exactly two callbacks: one for the source and one for
the destination solver to continue with after the transfer is complete.

The list of transfers describe the coupling among all solvers. This is
currently simply hard-coded to be the simple 0->1, i.e., mesh 0's
solution is prescribed to be transferred to mesh 1.

Once the callbacks are communicated among the solvers and all of them
are complete, all solvers also create their own list of transfers
(m_mytransfer) that is a subset of all prescribed transfers
(m_transfer). Only the transfers are included in m_mytransfer in which a
solver participates as either a source or a destination.

After this, we only transfer the initial conditions between 0 and 1, and
not within time stepping as of yet. The transfer is based on
m_mytransfer. In the future, multiple transfers will be requested
asynchronously from the library, issuing multiple pairs of calls to
setSourceTets() and setSourcePoints(). This already happens if multiple
transfers are configured, but the library is not ready for buffering
multiple transfers yet. Nevertheless, the above machinery is in place to
support multiple transfers, among multiple source and destination
solvers in an arbitrary fashion, e.g., single source to multiple
destinations, single destination from multiple sources. (This latter
will be necessary for, e.g., a solution transfer back from a coupled
multiple RBD solvers, holding different meshes, to the CFD solver,
computing a background flow on a single mesh.)