Put in dot product and matrix-vector multiply in conjugate gradients solver
Tested with a Laplacian as the matrix and some nonzero values for x and
b, computing the initial residual. Seems to work in both serial and
parallel, verified with octave, computing the norm of b and r=b-A*x.

The dot product is generic and parametrizable, while the matrix-vector
multiply is specific to computing the initial residual. This should be
okay, because there will be 3 dot products and only 1 matrix-vector
multiply during a conjugate gradients iteration. No tests yet.