Add parsing for mesh ... end block
This commit adds inciter parsing for mesh ... end blocks, that will be
used to configure and assign a mesh to a solver. All of Transport,
CompFlow, and MultiMat will be associated with a mesh to facilitate
coupling via mesh-to-mesh transfer. The configuration block for now can
parse and store the following:

  compflow
    depvar u    # note depvar is mandatory, labeling this solver
    # assign mesh file for this solver
    mesh filename "unitcube_4.5k.exo" end
  end

  compflow
    depvar v
    # assign mesh file for this solver, specifying a location and
    # orientation of this mesh relative to some coordinate system (this
    # latter is not yet specified)
    mesh
      filename "unitcube_31k.exo"
      location 0.0 1.0 2.0 end  # translate in x,y,z
      orientation 0.0 +90.2 -180.0 end  # rotate around x,y,z axes
    end
  end