Allow specifying input mesh on command line or multiple input meshes in control file
This commit puts in code to allow either specifying a single input mesh
on the command line or multiple input meshes (assigned to solvers) in
the control file. Either the former or the latter must be used, neither
or both is an error. The former allows for the original behavior of using a
single mesh and solving a system of systems on a single mesh, while the
latter allows using multiple solvers on different meshes and coupling
via mesh transfer.