Add mesh velocity to ALECG flux
This modifies the flux in ALECG for nonzero mesh velocity and updates
the mesh node coordinates with the mesh velocity. It does not recompute
the volumes yet.

This is tested by prescribing a simple translation for the mesh velocity
and running the Sod problem, which appears to be correct, because
running a Sod regression test passes on the diagnostics file. It diffs
on the field output because when the mesh moves we now save the new mesh
to a new file, as required by exodus.

Next is to recompute the nodal volumes and test the GCL.