Recompute mesh volumes in ALECG after ALE mesh movement
This commit puts in code to recompute the nodal mesh volumes after the
node coordinates have been advanced if ALE is enabled in ALECG.

The volumes have already been recomputed after AMR if mesh
refinement/derefinement happened in a time step. The same code is
invoked for recomputing the volumes after ALE. As a result, if both ALE
and AMR happened in a time step, the volumes are recomputed twice.  ALE
mesh movement happens after the linear solve and before AMR.

Recomputing the volumes twice if both ALE and AMR happened in a time
step is not necessarily the best, but for now this is kept simple. If
ALE enabled, the volumes need to be recomputed in every time step stage,
but AMR happens in the last time step stage only if the AMR is enabled
and if we hit the AMR frequency. To save on not recomputing the volumes
twice if both ALE and AMR happened in a time step stage would
complicated logic, so we keep this simple for now and not try to
optimize this. Will perhaps revisit this in the future.

The logic appears to work in serial and parallel, both with and without
ALE. However, AMR is not yet tested with ALECG.

Next is to test the geometric conservation law with ALE mesh movement.