Keep "sine" ALE mesh velocity domain-preserving
This removes the division from the sine wave from the "sine" mesh
velocity, used only for testing the symmetry BCs with moving ALE
boundaries. This results in preserving the total domain and mesh volume
and no moving boundaries, so it is easier to check and test known
results from test problems, such as the Sod shocktube for which now
there is also a regression test: testing the Sod solution with ALECG
computing the solution on a mesh with a prescribed velocity.