Modify ALECG logic to accommodate mesh velocity update during timestepping
This commit modifies the logic in ALECG to enable updating the mesh
velocity during timestepping in every time step stage. This a first step
towards assigning the fluid velocity as the mesh velocity to enable
Lagrangian mesh movement. To make that practical, we need some kind of
mesh smoothing so this commit also gets ready to solve a Laplacian for
smoothing the fluid velocity. This will be configured by the 'fluid'
mesh_velocity in the input file. Since we previously introduced simply
statically-prescribed sinusoidal mesh velocity for testing and we
already have regression tests exercising this, we want to keep that
going by now differentiating between 'dynamic' and 'static' mesh
velocity types. Static mesh velocities will only assign a mesh velocity
during setup and leave it alone during time stepping, while dynamic one,
e.g., 'fluid' will update the mesh velocity during time stepping.

Smoothing will also require the vorticity, which requires the fluid
velocity, so this commit also adds a way of querying the fluid velocity
from the system of PDEs. This should work for systems and so ALECG needs
to call down into the PDE class to query the fluid velocity. Obviously,
e.g., Transport should not even define a fluid velocity. At this point
this mechanism, querying the fluid velocity from multiple systems, is
rudimentary, because the last system that defines a fluid velocity will
be used as the "winner" fluid velocity which is then used for ALE mesh
motion. In the future, coupled solvers need to agree on who wins on the
fluid velocity question. This can probably be simply decided by the user
by labelling which system should define the fluid velocity used for ALE
if the systems are solved on the same mesh.

Although, the smoother's linear system solve is enabled, triggered by
'mesh_velocity fluid', the linear system is not yet correctly
initialized for a successful linear solve because boundary conditions on
the mesh velocity solve are still missing.

The conjugate gradients solver now only requires the max number of
iterations and the final tolerance in its solve() member function,
because those are not needed before (during construction and
initialization, computing the initial residuals). Hence the update on
the linear solver unit tests.