Conjugate gradients (CG) solver communicates BCs
Previously BCs on the linear solver were not consistent on different
chares. As a result chares that shared a row, corresponding to a mesh
node on which BC is set, contained different data in the matrix, leading
to a different solution in serial and parallel (but only if BCs were
applied).

Now the CG solver communicates the boundary conditions and this ensures
that all chares that contribute to the same node on which BC is applied
will be represented with the same data in the matrix after the BCs are
applied on different chares.

This is necessary because there can be nodes a chare contributes to but
does not apply BCs on. This happens if a node is in the node
communication map but not on the list of incoming BCs on a chare. To
have all chares share the same view on all BC nodes, we now send the
global node ids together with the Dirichlet BCs at which BCs are set to
those fellow chares that also contribute to those BC nodes. Only after
this communication step the chare applies the BCs on the matrix, which
then will correctly setup the BC rows that exist on multiple chares
(which now will be the same as the results of making the BCs consistent
across all chares that contribute.

Another change is that the CG solver now does not recompute the norm of
the residual in the beginning of a new iteration. This is not necessary,
since the end of the iteration also computes this, so now we simply
reuse the norm resulting from the end of the iteration. This eliminates
an all-to-all dot product. Unit tests do not change.

CSR::dirichlet() has also been simplified, leading to shorter, less
repetitive, code. No change in functionality and test results.

Removed the references from the CG algorithm steps documentation. It was
annoying that the function names showed up the editor search for
function names, I don't think it was terribly useful, and it was no
incorrect due to reorganization due to not recomputing the norm of the
residual in the beginning of a new iteration.

Removed the references to "DOF", using "ncomp" instead, to make this
consistent with the rest of the code.

Still kept some debug commented debug code until mesh velocity BCs are
finalized and generalized to arbitrarily oriented side sets and
surfaces.