ALE no longer diverges the mesh velocity on chare boundaries
Details:

The CG solver now contains a communication step at each solver
iteration. I still do not understand why this is necessary, but it
appears that in parallel, the solution in x may slightly different at
nodes shared by multiple chares and this can be fatal with more
iterations. In principle this step should not be necessary, but when it
is there, the solution is more of what we get in serial (modulo floating
point representation errors). I have tried several, seemingly and
logically correct, different strategies for communicating the solution
vector at the end of every iteration, for example keeping the one with
the largest chare id, etc., and the only one that appears to works best
is computing an average on shared nodes. The new communication step thus
simply adds up the solution on shared nodes and computes an average.
This works okay on nodes that are shared among multiple chares.

The type for storing mesh velocity BCs now do not hard-code the number
of scalar components to be 3. This is so the conjugate gradients (CG)
solver can accept arbitrary number of scalar components, as is the rest
of CG and CSR are already written.

The lambda, count(), previously only called from within
CSR::dirichlet(), now lives in Mesh/CommMap.[ch]pp, so it can be also
called (without repeating code) in ConjugateGradients::x().

CSR::dirichlet() now zeros the column as well as the row at Dirichlet BC
nodes.  This ensures faster convergence and more accurate solution. The
Sod problem breaks if we only zero the row (as before). This required a
base-line update for some unit tests.