Put in punt for generalizing inheriting meshvel BCs from symmetry BCs
This is practically a punt for generalizing the inheritance for mesh
velocity BCs from symmetry BCs. It is a punt because I have tried
various ways of using the boundary-point normals at symmetry BCs to
remove the normal component of the mesh velocity and none of them are
robust enough compared to directly setting Dirichlet BCs directly on the
components that can be invoked when the symmetry side sets are aligned
with coordinate boundaries. We may revisit this, but for now I put in
this punt which zeroes out the correct component of the Dirichlet BCs if
the symmetry BCs side set is aligned with a coordinate direction.