Replace sintrg search with exact match for assigning outvars
Previously, assigning output variables was based on a
std::string::find(), but that matched also for a substring, which may
lead to subtle bugs if output variables are named in a way that one is a
subtring of another one because it .find() matches the wrong output
variable and this assigns the wrong getvar().

This commit replaces the subtring search with an exact match search.
This allows finding only the exact match, so assigning the correct
output variable function. For example:

'vorticity' and 'x-vorticity' would previously be matched to whichever
was found first. Now the correct one is assigned.