Rename materialid_info::expect::description to a single word
As it turns out, and as I had surely known this but had to remind
myself, this field must be not contain a space, because then the
automated script, tools/extract_ctr_keywords.sh, that extracts
documentation from the executable, breaks.

This commit fixes the symptom, but ultimately it would be better to fix
the script to allow spaces.