Add symbc to ALE mesh velocity
This allows setting symmetry BCs to the ALE mesh velocity. This works by
applying the symmetry BCs on the mesh velocity after the Helmholtz
decomposition (and for now it only hooked up for Helmholtz-meshvel).

Example:

  ale
    dvcfl 1.0
    mesh_velocity helmholtz
    vortmult 1.0  # remove all vorticity
    maxit 1000
    tolerance 1.0e-3
    bc_sym
      sideset 1 2 3 4 5 6 end
    end
    bc_dirichlet
      sideset 1 2 3 4 5 6 end
    end
  end

The above will apply Dirichlet BCs on the Helmholtz linear solve for the
scalar potential, remove all vorticity from the mesh velocity, assign it
to the fluid velocity, and finally, removes the normal components of the
mesh velocity on side sets where bc_sym is specificed. Note that this
symbc is independent of symbcs for the fluid velocity.