Fix ALECG::m_veldiv and use v0/vn for the ALE mesh force
Fixing ALECG::m_veldiv entails storing the actual velocity divergence in
that array instead of its weak form. This is better since we now use it
for the mesh force as well and that needs the actual velocity divergence
and not its weak form, so this is also less confusing.

Using v0/vn instead of v0/v for computing the mesh force avoids a
problem in parallel where the current volumes are being recomputed while
the mesh force is also computed at the same time. Using the vn, the
volume at the previous time step, avoids this problem.