Put in parsing for moving sides for ALE
Now the following is correctly parsed and error checked:

inciter
  ...
  ale
    move
     sideset 2 end
     fntype position
     fn # t         x(t) y(t)        z(t)
          0         0    0           1
          0.0634665 0    0.0634239   0.997987
          0.126933  0    0.126592    0.991955
          0.1904    0    0.189251    0.981929
          ...
     end
    end
  end
  ...
end

The above specifies a single move ... end block within an ale ... end
block (multiple possible), which configures a list of side sets and a
user-defined function, given by discrete entries of the abscissa and a
triplet of ordinates, defining a 3D position in time. It is also
possible to select how the function should be interpreted by the code:
as a position, velocity, or acceleration. This function will be used to
move a list of side sets in time for ALE simulations.

Nothing is hooked up yet, this commit puts in parsing only.